Add files via upload
This is a FORWARD KINEMATICS simulation developed on SIMULINK & SIMSCAPE for the IIWA robot.
The maths used are very basic SCREW THEORY functions.
The simulation has a Dashboard to perform MANUAL motion of the robot JOINTS and alternatively an AUTO possibility for introducing a complete Trajectory for the magnitudes of the Joints. Besides a Safety button is provided to STOP the robot at any time.